👌 IMPROVE: Refactor color class to its own file
