+++ menus/display_menu.py
@@ -1,15 +1,4 @@
-from utils import BaseMenu
-
-
-class Color:
-    '''
-    Enum class for console colours
-    '''
-    # ANSI escape codes for colors
-    GREEN = '\033[92m'  # Green
-    YELLOW = '\033[93m'  # Yellow
-    RED = '\033[91m'  # Red
-    ENDC = '\033[0m'  # End Colour
+from utils import BaseMenu, Color
 
 
 class DisplayMenu(BaseMenu):
