+++ utils/__init__.py
@@ -1,2 +1,3 @@
 from .book import Book
 from .base_menu import BaseMenu
+from .color import Color
