+++ utils/color.py
@@ -1,0 +1,9 @@
+class Color:
+    '''
+    Enum class for console colours
+    '''
+    # ANSI escape codes for colors
+    GREEN = '\033[92m'  # Green
+    YELLOW = '\033[93m'  # Yellow
+    RED = '\033[91m'  # Red
+    ENDC = '\033[0m'  # End Colour
